@@ -13,85 +13,76 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/common": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "@ionic-native/background-mode": "^5.6.1",
-    "@ionic-native/core": "^5.1.0",
-    "@ionic-native/fcm": "^5.6.1",
-    "@ionic-native/firebase": "^5.0.0-beta.14",
-    "@ionic-native/geolocation": "^5.5.0",
-    "@ionic-native/in-app-browser": "^5.5.0",
-    "@ionic-native/instagram": "^5.6.0",
-    "@ionic-native/keyboard": "^5.5.1",
-    "@ionic-native/qr-scanner": "^5.4.0",
-    "@ionic-native/screen-orientation": "^5.4.0",
-    "@ionic-native/social-sharing": "^5.5.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "~8.1.2",
+    "@angular/core": "~8.1.2",
+    "@angular/forms": "~8.1.2",
+    "@angular/platform-browser": "~8.1.2",
+    "@angular/platform-browser-dynamic": "~8.1.2",
+    "@angular/router": "~8.1.2",
+    "@ionic-native/background-mode": "^5.20.0",
+    "@ionic-native/core": "^5.0.0",
+    "@ionic-native/fcm": "^5.21.2",
+    "@ionic-native/geolocation": "^5.20.0",
+    "@ionic-native/in-app-browser": "^5.20.0",
+    "@ionic-native/instagram": "^5.20.0",
+    "@ionic-native/keyboard": "^5.20.0",
+    "@ionic-native/qr-scanner": "^5.20.0",
+    "@ionic-native/social-sharing": "^5.20.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.1.0",
-    "@ionic/pro": "2.0.4",
-    "ajv": "^6.9.1",
-    "cordova-android": "8.0.0",
-    "cordova-instagram-plugin": "0.5.8",
-    "cordova-plugin-background-mode": "0.7.2",
-    "cordova-plugin-fcm-with-dependecy-updated": "2.4.0",
-    "cordova-plugin-geolocation": "4.0.1",
-    "cordova-plugin-inappbrowser": "3.0.0",
-    "cordova-plugin-ionic": "5.3.0",
-    "cordova-plugin-qrscanner": "3.0.1",
-    "cordova-plugin-screen-orientation": "3.0.1",
-    "cordova-plugin-x-socialsharing": "5.4.4",
-    "cordova-universal-clipboard": "git+https://github.com/Visigo/CordovaClipboard.git",
+    "@ionic/angular": "^4.7.1",
+    "build": "0.1.4",
+    "cordova-android": "8.1.0",
+    "cordova-plugin-background-mode": "^0.7.3",
+    "cordova-plugin-fcm-with-dependecy-updated": "^4.1.1",
+    "cordova-plugin-geolocation": "^4.0.2",
+    "cordova-plugin-inappbrowser": "^3.2.0",
+    "cordova-plugin-qrscanner": "^3.0.1",
+    "cordova-plugin-x-socialsharing": "^5.6.3",
     "core-js": "^2.5.4",
-    "es6-promise-plugin": "4.2.2",
-    "leaflet": "^1.4.0",
-    "ngx-qrcode2": "0.0.9",
-    "rxjs": "~6.3.3",
-    "start": "5.1.0",
-    "zone.js": "~0.8.29"
+    "es6-promise-plugin": "^4.2.2",
+    "ionic": "^5.4.16",
+    "rxjs": "~6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.12.3",
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular-devkit/core": "~7.2.3",
-    "@angular-devkit/schematics": "~7.2.3",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler": "^7.2.13",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
-    "@ionic/angular-toolkit": "~1.4.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/architect": "~0.801.2",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular-devkit/core": "~8.1.2",
+    "@angular-devkit/schematics": "~8.1.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/language-service": "~8.1.2",
+    "@ionic/angular-toolkit": "^2.1.1",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.14.2",
-    "codelyzer": "~4.5.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.0",
     "cordova-plugin-device": "^2.0.2",
-    "cordova-plugin-ionic-keyboard": "^2.1.3",
-    "cordova-plugin-ionic-webview": "^3.1.2",
+    "cordova-plugin-ionic-keyboard": "^2.2.0",
+    "cordova-plugin-ionic-webview": "^4.1.3",
     "cordova-plugin-splashscreen": "^5.0.2",
     "cordova-plugin-statusbar": "^2.4.2",
     "cordova-plugin-whitelist": "^1.3.3",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.0.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
-    "ts-node": "~8.0.0",
-    "tslint": "~5.12.0",
-    "typescript": "~3.1.6",
-    "webpack-dev-server": "^3.3.1"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
     "plugins": {
+      "cordova-plugin-geolocation": {},
       "cordova-plugin-whitelist": {},
       "cordova-plugin-statusbar": {},
       "cordova-plugin-device": {},
