@@ -3,8 +3,6 @@ import {ProfileService} from '../../services/profile.service';
 import {AlertService} from '../../services/alert.service';
 import {User} from '../../models/user';
 import {SocialSharing} from '@ionic-native/social-sharing/ngx';
-import {Instagram} from '@ionic-native/instagram/ngx';
-import {AlertController} from '@ionic/angular';
 
 @Component({
     selector: 'app-profile',
@@ -18,9 +16,7 @@ export class ProfileComponent {
 
     constructor(private profileService: ProfileService,
                 private alert: AlertService,
-                private socialSharing: SocialSharing,
-                private instagram: Instagram,
-                private alertController: AlertController) {
+                private socialSharing: SocialSharing) {
     }
 
     ionViewWillEnter() {
@@ -36,43 +32,15 @@ export class ProfileComponent {
         });
     }
 
-    async presentAlertMultipleButtons() {
-        const alert = await this.alertController.create({
-            header: 'Социальные сети',
-            buttons: [
-                {
-                    text: 'WhatsApp',
-                    handler: () => {
-                        this.shareWhatsApp();
-                    }
-                },
-                {
-                    text: 'Instagramm',
-                    handler: () => {
-                        this.shareInstagram();
-                    }
-                },
-                {
-                    text: 'Facebook',
-                    handler: () => {
-                        this.shareFacebook();
-                    }
-                }
-            ]
-        });
-
-        await alert.present();
-    }
-
-    async shareWhatsApp() {
+    presentAlertMultipleButtons() {
         this.profileService.getInvite().subscribe(res => {
             if (res.status === 1) {
-                this.socialSharing.shareViaWhatsApp(
+                this.socialSharing.share(
                     'Скачай приложение Изи Парк, введи промокод ' + res.msg + ' и получи 20 бонусов на аттракционы.',
                     null,
                     null)
-                    .then(() => {
-
+                    .then((ev) => {
+                        console.log('event', ev);
                     }).catch((e) => {
                     this.alert.onAlert('error', 'Ошибка отправления.');
                 });
@@ -84,33 +52,4 @@ export class ProfileComponent {
         });
     }
 
-    async shareFacebook() {
-        this.profileService.getInvite().subscribe(res => {
-            if (res.status === 1) {
-                this.socialSharing.shareViaFacebook(
-                    'Скачай приложение Изи Парк, введи промокод ' + res.msg + ' и получи 20 бонусов на аттракционы.',
-                    null,
-                    null)
-                    .then(() => {
-
-                    }).catch((e) => {
-                    this.alert.onAlert('error', 'Ошибка отправления.');
-                });
-            } else if (res.status === 0) {
-                this.alert.onAlert('error', res.msg);
-            }
-        }, error => {
-            this.alert.onAlert('error', 'Ошибка получения.');
-        });
-    }
-
-    shareInstagram() {
-        this.instagram.share('data:image/jpg;card.jpg', 'Заголовок')
-            .then(res => {
-                console.log(res);
-            })
-            .catch(err => {
-                this.alert.onAlert('error', 'Ошибка отправления.');
-            });
-    }
 }
