@@ -1,4 +1,4 @@
 export const environment = {
     production: true,
-    apiUrl: 'https://api.mobile.goldinnfish.com'
+    apiUrl: 'http://easypark.goldinnfish.com'
 };
